refactor(editor): Extract vim utilities into separate modules
- Split monolithic vim.ts into focused modules for types, clipboard, motions, and bracket helpers
- Create vimTypes.ts with shared TypeScript interfaces (VimAPI, CodeMirrorAdapter, VimOperatorArgs, etc.)
- Create vimClipboard.ts with createYankSystemOperator and createPasteSystemAction functions
- Create vimMotions.ts with moveByDisplayLinesMotion and moveToMatchingBracketMotion
- Create vimBracketHelpers.ts with bracket/fence/quote matching logic and improved type safety
- Add JSDoc comments to all exported functions for better documentation
- Improve type safety with SmartQuoteChar type and constants for regex patterns
